chore: move overlay and overlay navigation blocks to inner blocks
