+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/overlay/block.json
