+++ plugins/woocommerce-blocks/assets/js/blocks/product-filters/inner-blocks/overlay/settings.ts
